Add rest of the unused examples which generate output to verify
chapel.inline.sync.py is now test_sync.py, chapel.sfile.hello.ename.py is now
test_chapel_sfile.py, chapel.inline.for.py is now test_for.py and
chapel.inline.forall.py is now test_forall.py
