+++ docs/source/examples/test_chapel_sfile.py
@@ -4,5 +4,7 @@ from pych.extern import Chapel
 def hello_caller():
     return None
 
-if __name__ == "__main__":
+def test_sfile(capfd):
     hello_caller()
+    out, err = capfd.readouterr()
+    assert out == 'Hi Caller, I am Chapel, pleased to meet you.\n'
