+++ docs/source/examples/test_for.py
@@ -10,5 +10,8 @@ def ex_for(start=int, end=int):
     """
     return float
 
-if __name__ == "__main__":
+def test_for(capfd):
     ex_for(1, 1000)
+    out, err = capfd.readouterr()
+    for i in xrange(1, 1000):
+        assert "This is " + str(i) + "\n" in out
